requirements.txt
flask
flask-cors
pandas
python-docx
matplotlib
